+++ src/app/(views)/(public)/explore/page.tsx
@@ -1,5 +1,10 @@
+import { ChevronRightIcon } from "lucide-react";
+import Link from "next/link";
+
 import DetailsCard from "~/components/shared/details-card";
 import GridContainer from "~/components/shared/grid-container";
+import Marker from "~/components/shared/marker";
+import type { ValidMetadata } from "~/types-schema/types";
 import fetchApi from "~/utils/fetch";
 
 import LastWatched from "./last-watched";
@@ -8,8 +13,10 @@ export const dynamic = "force-dynamic";
 export const revalidate = 60 * 5; // Cache the page for 5 minutes, unless revalidated on updates
 
 export default async function ExplorePage() {
-  const catalogs = await fetchApi("/catalogs/valid");
-  const archives = await fetchApi("/archives/valid");
+  const catalogs = await fetchApi<ValidMetadata[]>("/catalogs/valid");
+  const archives = await fetchApi<Omit<ValidMetadata, "pageviews">[]>(
+    "/archives/valid"
+  );
 
   const catalogsData = catalogs?.data;
   const archivesData = archives?.data;
@@ -21,14 +28,12 @@ export default async function ExplorePage() {
       {/* Last watched */}
       <LastWatched />
       {/* Featured Catalogs */}
-      {catalogsData.length && ENABLE_FEATURED ? (
+      {catalogsData?.length && ENABLE_FEATURED ? (
         <section>
-          <h1 className="text-2xl font-semibold tracking-tight">
-            Featured catalogs
-          </h1>
+          <Title label="Featured catalogs" link="/explore/catalogs" />
           <div className="w-full pt-7">
             <GridContainer>
-              {catalogsData.slice(0, 4).map((catalog: any) => (
+              {catalogsData.slice(0, 4).map((catalog) => (
                 <DetailsCard
                   path={`/c/${catalog.id}`}
                   key={catalog.id}
@@ -41,14 +46,12 @@ export default async function ExplorePage() {
       ) : null}
 
       {/* Featured Archives */}
-      {archivesData.length && ENABLE_FEATURED ? (
+      {archivesData?.length && ENABLE_FEATURED ? (
         <section>
-          <h1 className="text-2xl font-semibold tracking-tight">
-            Featured archives
-          </h1>
+          <Title label="Featured archives" link="/explore/archives" />
           <div className="w-full pt-7">
             <GridContainer>
-              {archivesData.slice(0, 4).map((archive: any) => (
+              {archivesData.slice(0, 4).map((archive) => (
                 <DetailsCard
                   path={`/a/${archive.id}`}
                   key={archive.id}
@@ -60,5 +63,24 @@ export default async function ExplorePage() {
         </section>
       ) : null}
     </div>
+  );
+}
+
+function Title({ label, link }: { label: string; link?: string }) {
+  return (
+    <h1
+      className="h-7 text-2xl font-semibold tracking-tight text-primary flex items-center gap-2"
+      aria-label={label}
+    >
+      <Marker />
+      <div className="flex items-end gap-2">
+        <p>{label}</p>
+        {link ? (
+          <Link className="cursor-pointer" href={link}>
+            <ChevronRightIcon className="size-7 text-primary stroke-[3]" />
+          </Link>
+        ) : null}
+      </div>
+    </h1>
   );
 }