+++ src/components/shared/details-card.tsx
@@ -1,10 +1,18 @@
-import { EyeIcon } from "lucide-react";
+import { EyeIcon, VideoIcon } from "lucide-react";
 import Link from "next/link";
+
+import { ValidMetadata } from "~/types-schema/types";
 
 import ThumbnailCarousel from "./carousel-thumbnails";
 import JustTip from "./just-the-tip";
+import OverlayTip from "./overlay-tip";
 
-export default function DetailsCard({ pageData, path }: any) {
+interface DetailsCardProps {
+  pageData: ValidMetadata;
+  path: string;
+}
+
+export default function DetailsCard({ pageData, path }: DetailsCardProps) {
   return (
     <Link
       key={pageData?.id}
@@ -14,7 +22,10 @@ export default function DetailsCard({ pageData, path }: any) {
       <section className="aspect-video">
         <ThumbnailCarousel thumbnails={pageData.thumbnails} />
       </section>
-      <Pageview pageviews={pageData.pageviews} />
+      {pageData?.pageviews ? <Pageview pageviews={pageData.pageviews} /> : null}
+      {pageData?.totalVideos ? (
+        <TotalVideos totalVideos={pageData.totalVideos} />
+      ) : null}
 
       <div className="absolute inset-0 aspect-video bg-gradient-to-b from-transparent to-black/90"></div>
       <div className="p-4">
@@ -25,17 +36,31 @@ export default function DetailsCard({ pageData, path }: any) {
   );
 }
 
+function TotalVideos({ totalVideos }: { totalVideos: number }) {
+  return (
+    <JustTip label="Total videos">
+      <OverlayTip
+        id="total-videos"
+        className="flex gap-1 absolute top-2 left-2 items-center p-1 rounded-md z-20"
+      >
+        <p className="text-xs tracking-widest">{totalVideos}</p>
+        <VideoIcon className="size-3" />
+      </OverlayTip>
+    </JustTip>
+  );
+}
+
 function Pageview({ pageviews }: { pageviews: number }) {
   if (pageviews !== undefined) {
     return (
       <JustTip label="Unique views">
-        <div
+        <OverlayTip
           id="pageviews"
-          className="flex gap-1 absolute top-2 right-2 items-center p-1 bg-accent/70 rounded-md z-20"
+          className="flex gap-1 absolute top-2 right-2 items-center p-1 rounded-md z-20"
         >
-          <p className="text-xs">{pageviews}</p>
+          <p className="text-xs tracking-widest">{pageviews}</p>
           <EyeIcon className="size-3" />
-        </div>
+        </OverlayTip>
       </JustTip>
     );
   }